Refactored duplicate editions rendering code into helper func, initialized PublisherPage component
