+++ src/client/components/pages/entity/publication.js
@@ -19,10 +19,9 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 const React = require('react');
-const FontAwesome = require('react-fontawesome');
-const Button = require('react-bootstrap').Button;
-const Table = require('react-bootstrap').Table;
 const showEntityType = require('../../../helpers/entity').showEntityType;
+const showEntityEditions =
+	require('../../../helpers/entity').showEntityEditions;
 
 class PublicationPage extends React.Component {
 
@@ -36,60 +35,7 @@ class PublicationPage extends React.Component {
 
 	render() {
 		const {entity} = this.props;
-
-		return (
-			<div>
-				<h2>
-					Editions
-					<Button
-						bsStyle="success"
-						className="pull-right"
-						href={`/edition/create?publication=${entity.bbid}`}
-					>
-						<FontAwesome name="plus"/>
-						{'  Add Edition'}
-					</Button>
-				</h2>
-				<Table striped>
-					<thead>
-						<tr>
-							<th>Name</th>
-							<th>Release Date</th>
-						</tr>
-					</thead>
-					<tbody>
-					{entity.editions.map((edition, idx) => {
-						const editionName = edition.defaultAlias ?
-							edition.defaultAlias.name : '(unnamed)';
-						const editionComment = (edition.disambiguation &&
-						edition.disambiguation.comment) ?
-							` (${edition.disambiguation.comment})` : '';
-						const releaseEventDate = (edition.releaseEventSet &&
-						edition.releaseEventSet.releaseEvents &&
-						edition.releaseEventSet.releaseEvents[0]) &&
-							edition.releaseEventSet.releaseEvents[0].date;
-						return (
-							<tr
-								key={`${edition.bbid}${idx}`}
-							>
-								<td>
-									<a href={`/edition/${edition.bbid}`}>
-										{editionName}
-									</a>
-									<span className="text-muted">
-										{editionComment}
-									</span>
-								</td>
-								<td>
-									{releaseEventDate}
-								</td>
-							</tr>
-						);
-					})}
-					</tbody>
-				</Table>
-			</div>
-		);
+		return showEntityEditions(entity);
 	}
 }
 PublicationPage.displayName = 'PublicationPage';
